CORP decomposition of "point" Brier scores for clade 25A
As promised!

The motivation here is that I've been interested/excited lately about
CORP decomposition of scores. This is a version of the classic decomposition:

score = miscalibration - descrimination + uncertainty

but the particular CORP decomposition has some nice properties. See the paper
here for an explanation of why it should be preferred: https://arxiv.org/abs/2008.03033

The algorithm is implemented in the `{reliabilitydiag}` package. All I
did was take the targets pipeline and pull out the model data and subset
to 25A.

I expanded out the variant frequencies to successes/failures for
scoring.

However -- the `all_model_outputs_for_heatmap` target was too big for me
to load into RAM and run anything else. I had to subset to 25A and
convert UGA-multicast from samples to means in DuckDB
and save it to a parquet. I imagine you don't have this RAM limitation
on your machine and could just run it.

I really like the description in `?summary.reliabilitydiag`

>        ‘miscalibration’  a measure of miscalibration
                         (_how reliable is the prediction method?_),
                         smaller is better.
       ‘discrimination’  a measure of discrimination
                         (_how variable are the recalibrated predictions?_),
                         larger is better.
       ‘uncertainty’     the mean score of a constant prediction at the
                         value of the average observation.

There's more good stuff in there so take a look!

But my gist is something like "the HMLR and CADPH-CATaLog models have
more information content about clade 25A emergence than baseline after recalibration.
However, both are less reliable than baseline, with this miscalibration
severe enough in the HMLR model to reduce its score below baseline."

There are some fairly extensive limitations here:

* First and most importantly, this relies on the same approach as `brier_point`.
  It assumes "25A against all else" and scores on that, ignoring any
  multinomial structure.
* It's also using only the mean forecasts, not the uncertainty. So this
  captures the structural binomial uncertainty but not the model
  uncertainty (I imagine this might particularly penalize the HMLR).
     * It's possible we could rectify this by including multiple samples
       from each model? I'm not sure and need to read through some more
       to see if that's ok.
* Needs to run on complete cases. I dropped CATaMaran because it wasn't present
  for the whole beginning bit of 25A and the decomposition isn't
  comparable across subsets of data.

I promise I won't be offended if this doesn't make it in -- I really
did it for fun!